+++ Sources/EudiWalletKit/Services/StorageManager.swift
@@ -62,10 +62,10 @@ public class StorageManager: ObservableObject, @unchecked Sendable {
 	/// - Parameters:
 	///   - docs: An array of `WalletStorage.Document` objects to be refreshed.
 	///   - docStatus: The status of the documents.
-	private func refreshDocModels(_ docs: [WalletStorage.Document], docStatus: WalletStorage.DocumentStatus) async {
+	private func refreshDocModels(_ docs: [WalletStorage.Document], uiCulture: String?, docStatus: WalletStorage.DocumentStatus) async {
 		switch docStatus {
 		case .issued:
-			let models = docs.compactMap { Self.toClaimsModel(doc:$0, modelFactory: modelFactory) }
+			let models = docs.compactMap { Self.toClaimsModel(doc:$0, uiCulture: uiCulture, modelFactory: modelFactory) }
 			await MainActor.run { docModels = models }
 		case .deferred:
 			await MainActor.run { deferredDocuments = docs }
@@ -74,18 +74,18 @@ public class StorageManager: ObservableObject, @unchecked Sendable {
 		}
 	}
 
-	private func refreshDocModel(_ doc: WalletStorage.Document, docStatus: WalletStorage.DocumentStatus) async {
+	private func refreshDocModel(_ doc: WalletStorage.Document, uiCulture: String?, docStatus: WalletStorage.DocumentStatus) async {
 		if docStatus == .issued && docModels.first(where: { $0.id == doc.id}) == nil ||
 			docStatus == .deferred && deferredDocuments.first(where: { $0.id == doc.id}) == nil ||
 			docStatus == .pending && pendingDocuments.first(where: { $0.id == doc.id}) == nil {
-			_ = await appendDocModel(doc)
+			_ = await appendDocModel(doc, uiCulture: uiCulture)
 		}
 	}
 	
-	@discardableResult func appendDocModel(_ doc: WalletStorage.Document) async -> (any DocClaimsDecodable)? {
+	@discardableResult func appendDocModel(_ doc: WalletStorage.Document, uiCulture: String?) async -> (any DocClaimsDecodable)? {
 		switch doc.status {
 		case .issued:
-			let mdoc: (any DocClaimsDecodable)? = Self.toClaimsModel(doc: doc, modelFactory: modelFactory)
+			let mdoc: (any DocClaimsDecodable)? = Self.toClaimsModel(doc: doc, uiCulture: uiCulture, modelFactory: modelFactory)
 			if let mdoc { await MainActor.run { docModels.append(mdoc) } }
 			return mdoc
 		case .deferred:
@@ -113,17 +113,17 @@ public class StorageManager: ObservableObject, @unchecked Sendable {
 	///   - modelFactory: An optional factory conforming to `MdocModelFactory` to create the model. Defaults to `nil`.
 	///
 	/// - Returns: An optional `DocClaimsDecodable` model created from the given document.
-	public static func toClaimsModel(doc: WalletStorage.Document, modelFactory: (any DocClaimsDecodableFactory)? = nil) -> (any DocClaimsDecodable)? {
+	public static func toClaimsModel(doc: WalletStorage.Document, uiCulture: String?, modelFactory: (any DocClaimsDecodableFactory)? = nil) -> (any DocClaimsDecodable)? {
 		switch doc.docDataFormat {
-		case .cbor:	toCborMdocModel(doc: doc, modelFactory: modelFactory)
-		case .sdjwt: toSdJwtDocModel(doc: doc)
+		case .cbor:	toCborMdocModel(doc: doc, uiCulture: uiCulture, modelFactory: modelFactory)
+		case .sdjwt: toSdJwtDocModel(doc: doc, uiCulture: uiCulture)
 		}
 	}
 
-	public static func toCborMdocModel(doc: WalletStorage.Document, modelFactory: (any DocClaimsDecodableFactory)? = nil) -> (any DocClaimsDecodable)? {
+	public static func toCborMdocModel(doc: WalletStorage.Document, uiCulture: String?, modelFactory: (any DocClaimsDecodableFactory)? = nil) -> (any DocClaimsDecodable)? {
 		guard let (d, _, _) = doc.getDataForTransfer(), let iss = IssuerSigned(data: d.1.bytes) else { return nil }
 		let docMetadata: DocMetadata? = DocMetadata(from: doc.metadata)
-		let md = docMetadata?.getCborClaimMetadata()
+		let md = docMetadata?.getCborClaimMetadata(uiCulture: uiCulture)
 		var retModel: (any DocClaimsDecodable)? = modelFactory?.makeClaimsDecodableFromCbor(id: d.0, createdAt: doc.createdAt, issuerSigned: iss, displayName: md?.0, claimDisplayNames: md?.1, mandatoryClaims: md?.2, claimValueTypes: md?.3)
 		if retModel == nil {
 			let defModel: (any DocClaimsDecodable)? = switch doc.docType {
@@ -136,30 +136,32 @@ public class StorageManager: ObservableObject, @unchecked Sendable {
 		return retModel
 	}
 
-	public static func toSdJwtDocModel(doc: WalletStorage.Document, modelFactory: (any DocClaimsDecodableFactory)? = nil) -> (any DocClaimsDecodable)? {
+	public static func toSdJwtDocModel(doc: WalletStorage.Document, uiCulture: String?, modelFactory: (any DocClaimsDecodableFactory)? = nil) -> (any DocClaimsDecodable)? {
 		var docClaims = [DocClaim]()
 		let docMetadata: DocMetadata? = DocMetadata(from: doc.metadata)
-		let md = docMetadata?.getFlatClaimMetadata()
+		let md = docMetadata?.getFlatClaimMetadata(uiCulture: uiCulture)
 		let parser = CompactParser()
 		guard let serString = String(data: doc.data, encoding: .utf8), let sdJwt = try? parser.getSignedSdJwt(serialisedString: serString), let result = try? sdJwt.recreateClaims() else { return nil }
 		if let cs = result.recreatedClaims.toClaimsArray(md?.1, md?.2, md?.3)?.0 { docClaims.append(contentsOf: cs) }
 		var type = docClaims.first(where: { $0.name == "vct"})?.stringValue
 		if type == nil || type!.isEmpty { type = docClaims.first(where: { $0.name == "evidence"})?.children?.first(where: { $0.name == "type"})?.stringValue }
-		return GenericMdocModel(id: doc.id, createdAt: doc.createdAt, docType: doc.docType ?? type, displayName: docMetadata?.displayName, docClaims: docClaims, docDataFormat: .sdjwt)
+		return GenericMdocModel(id: doc.id, createdAt: doc.createdAt, docType: doc.docType ?? type, displayName: docMetadata?.getDisplayName(uiCulture), docClaims: docClaims, docDataFormat: .sdjwt)
 	}
 
-	public func getDocIdsToTypes() -> [String: (String, String?)] {
-		Dictionary(uniqueKeysWithValues: docModels.filter { $0.docType != nil}.map { m in (m.id, (m.docType!, m.displayName) ) })
+	public func getDocIdsToTypes() -> [String: (String, DocDataFormat, String?)] {
+		Dictionary(uniqueKeysWithValues: docModels.filter { $0.docType != nil}.map { m in (m.id, (m.docType!, m.docDataFormat, m.displayName) ) })
 	}
 	
 	/// Load documents from storage
 	///
 	/// Internally sets the ``docModels``,  ``mdlModel``, ``pidModel`` variables
 	/// - Returns: An array of ``WalletStorage.Document`` objects
-	@discardableResult public func loadDocuments(status: WalletStorage.DocumentStatus) async throws -> [WalletStorage.Document]?  {
+	@discardableResult public func loadDocuments(status: WalletStorage.DocumentStatus, uiCulture: String?) async throws -> [WalletStorage.Document]?  {
 		do {
 			guard let docs = try await storageService.loadDocuments(status: status) else { return nil }
-			await refreshDocModels(docs, docStatus: status)
+			let docs2 = docs.map { d in WalletStorage.Document(id: d.id, docType: d.docType, docDataFormat: d.docDataFormat, data: d.data, secureAreaName: d.secureAreaName,
+			 createdAt: d.createdAt, modifiedAt: d.modifiedAt, metadata: d.metadata, displayName: d.getDisplayName(uiCulture), status: d.status)   }
+			await refreshDocModels(docs2, uiCulture: uiCulture, docStatus: status)
 			await refreshPublishedVars()
 			return docs
 		} catch {
@@ -173,10 +175,10 @@ public class StorageManager: ObservableObject, @unchecked Sendable {
 	/// - Returns: A ``WalletStorage.Document`` object
 	/// - Parameter id: Identifier of document to load
 	/// - Parameter status: Status of document to load
-	@discardableResult public func loadDocument(id: String, status: DocumentStatus) async throws -> WalletStorage.Document?  {
+	@discardableResult public func loadDocument(id: String, uiCulture: String?, status: DocumentStatus) async throws -> WalletStorage.Document?  {
 		do {
 			guard let doc = try await storageService.loadDocument(id: id, status: status) else { return nil }
-			await refreshDocModel(doc, docStatus: status)
+			await refreshDocModel(doc, uiCulture: uiCulture, docStatus: status)
 			await refreshPublishedVars()
 			return doc
 		} catch {